Updated entities tipes to improve data types size

+++ backend/go/src/model/map_note_placement.go
@@ -5,7 +5,7 @@ type MapNotePlacement struct {
 	ID           ULID
 	MapNoteID    ULID
 	TargetNoteID ULID
-	X            float64
-	Y            float64
+	X            uint8
+	Y            uint8
 	AuditFields  AuditFields
 }
